+++ backend/src/features/financial_month/dao.ts
@@ -3,7 +3,6 @@ import { z } from 'zod'
 import type {
   FinancialMonth, FinancialMonthData, Months,
 } from '@/domains/financial_month'
-import { getPeriodByFinancialMonth } from '@/domains/financial_month/logic'
 import type dayjs from '@/logic/dayjs'
 import { condition, every } from '@/logic/queryBuilder/conditionTree'
 import { d1 } from '@/logic/queryBuilder/d1'
@@ -17,25 +16,7 @@ const FinancialMonthRecord = z.object({
   ended_at: z.number(),
 })
 
-type FinancialMonthRecord = z.infer<typeof FinancialMonthRecord>
-
-const makeRecord = ({
-  id, userId, financialYear, month,
-}: FinancialMonth): FinancialMonthRecord => {
-  const { start, end } = getPeriodByFinancialMonth({
-    financialYear,
-    month,
-  })
-
-  return {
-    id,
-    user_id: userId,
-    financial_year: financialYear,
-    month,
-    started_at: start.valueOf(),
-    ended_at: end.valueOf(),
-  }
-}
+export type FinancialMonthRecord = z.infer<typeof FinancialMonthRecord>
 
 const makeEntity = ({
   id, user_id, financial_year, month,
@@ -45,27 +26,6 @@ const makeEntity = ({
   financialYear: financial_year,
   month: month as Months,
 })
-
-// TODO: ビジネスロジック上12件単位で突っ込むのが当たり前なので、そうしたい
-// TODO: ここで勤務日数も初期化したい、というかこのメソッド自体financial_yearにいるべきかも
-export const insertFinancialMonth = (db: D1Database):
-(item: FinancialMonth) => Promise<FinancialMonth> =>
-  async (entity) => {
-    const stmt = 'INSERT INTO financial_months VALUES (?1,?2,?3,?4,?5,?6)'
-
-    const record = makeRecord(entity)
-
-    await db.prepare(stmt).bind(
-      record.id,
-      record.user_id,
-      record.financial_year,
-      record.month,
-      record.started_at,
-      record.ended_at,
-    ).run()
-
-    return entity
-  }
 
 export const findFinancialMonthsByYear = (db: D1Database):
 (userId: string, financialYear: number) => Promise<FinancialMonth[]> =>
